Minor fix to multiseg form to ensure base directory is correctly parsed. Minor fix to new simulation form to allow for streamnetwork span to be a floating point rather than integer value.

+++ GUI/GUI.AQUATOX/NewSimForm.cs
@@ -145,7 +145,7 @@ namespace GUI.AQUATOX
                 return;
             }
 
-            if ((!Int32.TryParse(EndCOMIDBox.Text, out int endCOMID)) && (!Int32.TryParse(spanBox.Text, out int Span)))
+            if ((!Int32.TryParse(EndCOMIDBox.Text, out int endCOMID)) && (!Double.TryParse(spanBox.Text, out double Span)))
             {
                 MessageBox.Show("Please either specify an up-river span, enter an up-river COMID in the endCOMID box, or right-click on a stream segment to select an upstream end point.");
                 return;
